+++ src/service/hotkeys.rs
@@ -2,7 +2,9 @@ use seelen_core::state::{shortcuts::SluHotkeyAction, Settings};
 use slu_ipc::{messages::AppMessage, AppIpc};
 use win_hotkeys::{error::WHKError, events::KeyboardInputEvent, Hotkey, HotkeyManager, VKey};
 
-use crate::{app_management::kill_seelen_ui_processes, error::Result, exit, log_error};
+use crate::{
+    app_management::kill_all_seelen_ui_processes, error::Result, exit, get_async_handler, log_error,
+};
 
 pub fn start_app_shortcuts(settings: &Settings) -> Result<()> {
     if let Err(err) = HotkeyManager::start_keyboard_capturing() {
@@ -35,24 +37,23 @@ pub fn start_app_shortcuts(settings: &Settings) -> Result<()> {
         }
 
         let action = slu_hotkey.action;
-        let tokio_handle = tokio::runtime::Handle::current();
 
         let hotkey = Hotkey::from_keys(vkeys).action(move || {
             log::trace!("Hotkey triggered: {action:?}");
             match action {
                 SluHotkeyAction::MiscForceRestart => {
-                    log_error!(kill_seelen_ui_processes());
+                    log_error!(kill_all_seelen_ui_processes());
                 }
                 SluHotkeyAction::MiscForceQuit => {
                     crate::EXITING.store(true, std::sync::atomic::Ordering::SeqCst);
-                    log_error!(kill_seelen_ui_processes());
+                    log_error!(kill_all_seelen_ui_processes());
                     exit(0);
                 }
                 _ => {}
             }
 
             if let Some(command) = hotkey_action_to_cli_command(action) {
-                tokio_handle.spawn(async move {
+                get_async_handler().spawn(async move {
                     log_error!(AppIpc::send(AppMessage::Cli(command)).await);
                 });
             }