+++ src/service/windows_api/mod.rs
@@ -2,13 +2,8 @@ pub mod app_bar;
 pub mod com;
 pub mod iterator;
 
-use std::{
-    ffi::OsString,
-    os::windows::ffi::OsStringExt,
-    path::{Path, PathBuf},
-};
+use std::{ffi::OsString, os::windows::ffi::OsStringExt, path::PathBuf};
 
-use com::Com;
 use windows::Win32::{
     Foundation::{HANDLE, HWND, LUID},
     Security::{
@@ -16,13 +11,16 @@ use windows::Win32::{
         TOKEN_ADJUST_PRIVILEGES, TOKEN_PRIVILEGES, TOKEN_QUERY,
     },
     System::{
-        Com::IPersistFile,
         Console::GetConsoleWindow,
-        Threading::{AttachThreadInput, GetCurrentProcess, GetCurrentThreadId, OpenProcessToken},
+        RemoteDesktop::ProcessIdToSessionId,
+        Threading::{
+            AttachThreadInput, GetCurrentProcess, GetCurrentProcessId, GetCurrentThreadId,
+            OpenProcessToken,
+        },
     },
     UI::{
         HiDpi::{SetProcessDpiAwarenessContext, DPI_AWARENESS_CONTEXT_PER_MONITOR_AWARE_V2},
-        Shell::{IShellLinkW, SHGetKnownFolderPath, ShellLink, KF_FLAG_DEFAULT},
+        Shell::{SHGetKnownFolderPath, KF_FLAG_DEFAULT},
         WindowsAndMessaging::{
             BringWindowToTop, FindWindowW, GetClassNameW, GetForegroundWindow, GetWindowTextW,
             GetWindowThreadProcessId, IsIconic, SetWindowPos, ShowWindow, ShowWindowAsync,
@@ -30,7 +28,7 @@ use windows::Win32::{
         },
     },
 };
-use windows_core::{Interface, PCWSTR};
+use windows_core::PCWSTR;
 
 use crate::{
     error::{Result, WindowsResultExt},
@@ -144,8 +142,20 @@ impl WindowsApi {
         unsafe { GetCurrentProcess() }
     }
 
+    pub fn current_process_id() -> u32 {
+        unsafe { GetCurrentProcessId() }
+    }
+
     pub fn current_thread_id() -> u32 {
         unsafe { GetCurrentThreadId() }
+    }
+
+    pub fn current_session_id() -> u32 {
+        let process_id = Self::current_process_id();
+        let mut session_id = 0;
+        // this should never fail for own process
+        unsafe { ProcessIdToSessionId(process_id, &mut session_id).expect("Can't get session id") };
+        session_id
     }
 
     pub fn open_current_process_token() -> Result<HANDLE> {
@@ -181,26 +191,6 @@ impl WindowsApi {
 
         unsafe { AdjustTokenPrivileges(token_handle, false, Some(&tkp), 0, None, None)? };
         Ok(())
-    }
-
-    pub fn create_temp_shortcut(program: &Path, args: &str) -> Result<PathBuf> {
-        Com::run_with_context(|| unsafe {
-            let shell_link: IShellLinkW = Com::create_instance(&ShellLink)?;
-
-            let program = WindowsString::from_os_string(program.as_os_str());
-            shell_link.SetPath(program.as_pcwstr())?;
-
-            let arguments = WindowsString::from_str(args);
-            shell_link.SetArguments(arguments.as_pcwstr())?;
-
-            let temp_dir = std::env::temp_dir();
-            let lnk_path = temp_dir.join(format!("{}.lnk", uuid::Uuid::new_v4()));
-            let lnk_path_wide = WindowsString::from_os_string(lnk_path.as_os_str());
-
-            let persist_file: IPersistFile = shell_link.cast()?;
-            persist_file.Save(lnk_path_wide.as_pcwstr(), true)?;
-            Ok(lnk_path)
-        })
     }
 
     // change to some crate like dirs to allow multiple platforms