+++ changelog.md
@@ -26,6 +26,8 @@
 - window manager moving windows being currently dragged by the user.
 - power menu on multi-monitor setups.
 - showing paused message on no video wallpapers.
+- app not working on local accounts.
+- app and service not running if already running on another session.
 
 ## [2.4.11]